+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Breaking Changes
+
+- Remove migration of options from versions before 4.0.0.
+
 ## 5.6.0 (2025-02-17)
 
 ### Enhancements